+++ HISTORY.rst
@@ -1,8 +1,13 @@
 Version History
 ===============
 
-Unreleased
-----------
+v0.5.2
+------
+
+New Features
+^^^^^^^^^^^^
+* ``average_over_dims`` added into ``average.core`` and ``average.ops``
+* New ``core.average.average_shape`` + ``core.subset.subset_create_weight_masks``. Depends on `xESMF` >= 0.5.2, which is a new optional dependency.
 
 Bug Fixes
 ^^^^^^^^^
@@ -18,8 +23,8 @@ Other Changes
 * Added a ``dev`` recipe for pip installations (`pip install clisops[dev]`).
 * Updated pre-commit and pre-commit hooks to newest versions.
 * Migrated linux-based integration builds to GitHub CI.
-* New ``core.average`` submodule with ``average_shape`` + ``core.subset_create_weight_masks``. Depends on `xESMF` >= 0.5.2, which is a new optional dependency.
 * Added functionality to ``core.subset.create_mask`` so it can accept ``GeoDataFrames`` with non-integer indexes.
+* ``clisops.utils.file_namers`` adjusted to allow values to be overwritten and extras to be added to the end before the file extension.
 
 
 v0.5.1 (2021-01-11)